Move ext. auth buttons to the separate component
